requirements.txt
distribute==0.6.35
Crypto
Django==1.4.3
Markdown==2.2.1
PyYAML==3.10
South==0.7.6
django-filter==0.5.4
django-rest==0.0.1
djangorestframework==2.1.15
mimeparse==0.1.3
pep8==1.4
pycrypto==2.6
python-dateutil==1.5
wsgiref==0.1.2
python-memcached==1.48
celery==3.0.15
django-celery==3.0.11
psycopg2==2.4.5
django-admin-bootstrapped==0.3.2
